+++ tests/unit/test_likelihoods.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pybamm
 import pytest
 
 import pybop
@@ -13,7 +14,8 @@ class TestLikelihoods:
 
     @pytest.fixture
     def model(self, ground_truth):
-        model = pybop.lithium_ion.SPM()
+        solver = pybamm.IDAKLUSolver()
+        model = pybop.lithium_ion.SPM(solver=solver)
         model.parameter_set.update(
             {
                 "Negative electrode active material volume fraction": ground_truth,
@@ -27,10 +29,12 @@ class TestLikelihoods:
 
     @pytest.fixture
     def parameters(self):
-        return pybop.Parameter(
-            "Negative electrode active material volume fraction",
-            prior=pybop.Gaussian(0.5, 0.01),
-            bounds=[0.375, 0.625],
+        return pybop.Parameters(
+            pybop.Parameter(
+                "Negative electrode active material volume fraction",
+                prior=pybop.Gaussian(0.5, 0.01),
+                bounds=[0.375, 0.625],
+            )
         )
 
     @pytest.fixture
@@ -204,3 +208,31 @@ class TestLikelihoods:
         np.testing.assert_allclose(
             grad, scaled_likelihood([0.6], calculate_grad=True)[1]
         )
+
+    @pytest.mark.parametrize(
+        "likelihood_cls",
+        [
+            pybop.GaussianLogLikelihoodKnownSigma,
+            pybop.JaxGaussianLogLikelihoodKnownSigma,
+        ],
+    )
+    def test_fisher_matrix(
+        self, likelihood_cls, one_signal_problem, model, dataset, parameters
+    ):
+        likelihood = likelihood_cls(one_signal_problem, sigma0=1e-3)
+        fisher = likelihood.observed_fisher([0.5])
+        assert isinstance(fisher, np.ndarray)
+
+        # Test fisher does not compute for non-gradient available parameters
+
+        parameters.add(
+            pybop.Parameter(
+                "Negative particle radius [m]",
+                prior=pybop.Gaussian(6e-06, 0.1e-6),
+                bounds=[1e-6, 9e-6],
+            ),
+        )
+        problem = pybop.FittingProblem(model, parameters, dataset)
+        likelihood_non_grad = likelihood_cls(problem, sigma0=1e-3)
+        fisher = likelihood_non_grad.observed_fisher([0.5, 5e-06])
+        assert fisher is None